% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaaroundus.R
\name{getcatchdata}
\alias{getcatchdata}
\title{Get catch data for a region as a dataframe or stacked area chart}
\usage{
getcatchdata(region, id, measure = "tonnage", dimension = "taxon",
  limit = 10, chart = F)
}
\description{
Get catch data for a region as a dataframe or stacked area chart
}

