% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getcelldata}
\alias{getcelldata}
\title{Get a dataframe with catch data for a given list of cells and year}
\usage{
getcelldata(year = 2010, cells, ...)
}
\arguments{
\item{year}{(integer/numeric) year of data. Default: 2010}

\item{cells}{(vector/list) list of cell IDs}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
data frame with catch data for the requested cells and year
}
\description{
Get a dataframe with catch data for a given list of cells and year
}
\examples{
getcelldata(2004, cells = 89568)
getcelldata(2008, cells = c(89568, 89569))
getcelldata(2011, cells = c(89568, 90288, 89569))
}
