% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaaroundus.R
\name{getregionmap}
\alias{getregionmap}
\title{Get a map of the region specified (TODO)}
\usage{
getregionmap(region, id)
}
\arguments{
\item{region}{region type}

\item{id}{region id}
}
\value{
map of the region
}
\description{
Get a map of the region specified (TODO)
}
\examples{
getregionmap("eez", 76)
}

