% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listregions.R
\name{listregions}
\alias{listregions}
\title{List available regions for a region type}
\usage{
listregions(region, env = "prod")
}
\arguments{
\item{region}{region type}

\item{env}{API environment
Default: "prod"}
}
\value{
a data frame with region ids and title(s)
}
\description{
List available regions for a region type
}
\examples{
listregions("eez")
}

