% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionmap.R
\name{regionmap}
\alias{regionmap}
\title{Get a map of the region specified}
\usage{
regionmap(region, id)
}
\arguments{
\item{region}{(character) region type. one of: eez, lme, rfmo, highseas,
fao, eez-bordering, fishing-entity, taxon, or global}

\item{id}{(character/integer/numeric) region id. depends on what
\code{region} is specified. See \code{\link{regions}} for details}
}
\value{
map of the region
}
\description{
Get a map of the region specified
}
\examples{
\dontrun{
regionmap("eez")
regionmap("eez", 76)

# a different region type
regionmap("lme", 23)
}
}
