% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\name{regions}
\alias{regions}
\title{Regions details}
\description{
Regions details
}
\details{
Region options
\itemize{
\item eez
\item lme
\item rfmo
\item highseas
\item fao
\item eez-bordering
\item fishing-entity
\item taxon
\item global
}
}
