% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaaroundus.R
\docType{package}
\name{rseaaroundus}
\alias{rseaaroundus}
\alias{rseaaroundus-package}
\title{R library for Sea Around Us}
\description{
: Access Sea Around Us catch data and view it as data frames
or stacked area charts.
}
\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}

Robert Scott Reis \email{reis.robert.s@gmail.com}
}
\keyword{package}
