% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaaroundus.R
\docType{package}
\name{seaaroundus-package}
\alias{seaaroundus-package}
\title{R library for Sea Around Us}
\description{
: Access Sea Around Us catch data and view it as data frames
or stacked area charts.
}
\section{Catch reconstruction documentation}{

\url{http://www.seaaroundus.org/catch-reconstruction-and-allocation-methods/}
\url{https://s3-us-west-2.amazonaws.com/sau-methods-docs/reconstruction-allocation/Methods-Catch-tab-Apr-29-2016.pdf}
}

\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}

Robert Scott Reis \email{reis.robert.s@gmail.com}
}
\keyword{package}
