% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchdata.R
\name{catchdata}
\alias{catchdata}
\title{Get catch data for a region as a dataframe}
\usage{
catchdata(region, id, ...)
}
\arguments{
\item{region}{(character) region of the data. Only one of "eez", "lme", "rfmo", "meow" will return data.}

\item{id}{(integer/numeric) region id. depends on what
\code{region} is specified. See \code{\link{listregions}} for details .}

\item{...}{curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}.}
}
\value{
data frame with catch data for the requested
region over the period 1950-2014.
}
\description{
Get catch data for a region as a dataframe
}
\examples{
catchdata("eez", 76)
head(catchdata("eez", 76))
catchdata("rfmo", 1)

}
