% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchdata.R
\name{catchdata}
\alias{catchdata}
\title{Get catch data for a region as a dataframe or stacked area chart}
\usage{
catchdata(region, id, measure = "tonnage", dimension = "taxon",
  limit = 10, chart = FALSE)
}
\arguments{
\item{region}{region type}

\item{id}{region id}

\item{measure}{measure of the data
Default: "tonnage"}

\item{dimension}{dimension data is bucketed on
Default: "taxon"}

\item{limit}{number of buckets of data plus one for "others"
Default: 10}

\item{chart}{boolean to return a chart versus a data frame
Default: \code{FALSE}}
}
\value{
data frame (or chart) with catch data for the requested region over time
}
\description{
Get catch data for a region as a dataframe or stacked area chart
}
\examples{
catchdata("eez", 76)
catchdata("eez", 76, measure="value", dimension="reporting-status")
\dontrun{
catchdata("eez", 76, chart=TRUE)
}
}
