% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getcelldata}
\alias{getcelldata}
\title{Get a dataframe with catch data for a given list of cells and year}
\usage{
getcelldata(year = 2010, cells)
}
\arguments{
\item{year}{year of data
Default: 2010}

\item{cells}{list of cell IDs}
}
\value{
data frame with catch data for the requested cells and year
}
\description{
Get a dataframe with catch data for a given list of cells and year
}
\examples{
getcelldata(2005, c(89568,90288,89569))
}
