% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{getcells}
\alias{getcells}
\title{Get list of cells in a given shape}
\usage{
getcells(shape)
}
\arguments{
\item{shape}{WKT representation of SRID 4326 polygon/multipolygon}
}
\value{
list of cell ids
}
\description{
Get list of cells in a given shape
}
\examples{
getcells("POLYGON ((-48.177685950413291 15.842380165289299,-48.177685950413291 15.842380165289299,
-54.964876033057919 28.964280991735578,-35.960743801652967 27.606842975206646,-48.177685950413291 
15.842380165289299))")
}
