% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listregions.R
\name{listregions}
\alias{listregions}
\title{List available regions for a region type}
\usage{
listregions(region, ...)
}
\arguments{
\item{region}{(character) region type}

\item{...}{curl options passed on to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
a data frame with columns:
\itemize{
\item region title
\item region ID
}
}
\description{
List available regions for a region type
}
\examples{
listregions(region = "eez")
}
