% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listregions.R
\name{listregions}
\alias{listregions}
\title{List available regions for a region type}
\usage{
listregions(region, ...)
}
\arguments{
\item{region}{region type}

\item{...}{curl options passed on to \code{\link[httr:GET]{httr::GET()}}}
}
\value{
a data frame with region ids and title(s)
}
\description{
List available regions for a region type
}
\examples{
listregions("eez")
}
