% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaaroundus.R
\name{listregions}
\alias{listregions}
\title{List available regions for a region type}
\usage{
listregions(region)
}
\arguments{
\item{region}{region type}
}
\value{
a data frame with region ids and title(s)
}
\description{
List available regions for a region type
}
\examples{
listregions("eez")
}

