% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionmap.R
\name{regionmap}
\alias{regionmap}
\title{Get a map of the region specified}
\usage{
regionmap(region, id)
}
\arguments{
\item{region}{region type}

\item{id}{region id}
}
\value{
map of the region
}
\description{
Get a map of the region specified
}
\examples{
regionmap("eez")
regionmap("eez", 76)
}

